chatbot/list.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

# Load your API key from .env file
load_dotenv()

# Configure the API key
genai.configure(api_key=os.environ["GOOGLE_API_KEY"])

# List available models
for m in genai.list_models():
    if 'generateContent' in m.supported_generation_methods:
        print(m.name)