requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
google-genai
langchain-google-genai
google-generativeai
bs4
pypdf
chromadb
faiss-cpu
arxiv